+++ extensions/naver-search/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Naver Search Changelog
 
+## [Add Feature] - {PR_MERGE_DATE}
+* Search Selected Word: Search for the selected word on the [Naver](https://naver.com).
+
 ## [Bugfix] - 2024-05-17
 
 * Fix Search Korean-English Dictionary `Not Found` error: Update Naver dictionary API endpoint.